requirements.txt
Django>=5.2.1
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-extensions>=3.2.0
Pillow>=10.0.0
